+++ src/tables/mod.rs
@@ -29,6 +29,7 @@ pub mod head;
 pub mod hhea;
 pub mod hmtx;
 pub mod hvar;
+pub mod jstf;
 pub mod kern;
 pub mod loca;
 pub mod ltsh;